requirements.txt
absl-py==1.4.0
blessings==1.7
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
einops==0.6.1
fonttools==4.43.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
imageio==2.9.0
joblib==1.3.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
path==16.7.1
Pillow==10.2.0
prettytable==3.8.0
progressbar2==4.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
torch==1.13.1
torchvision==0.2.2.post3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
wandb==0.16.2
wcwidth==0.2.6
Werkzeug==2.3.8
